// backend/src/search/dto/search-profile-list.dto.ts
    export class SearchProfileListDto {
        id: number;
        career: number;
        pay: number;
        startDate: number;
        possibleArea: string;
        license: string;
        keywords: string;
        notice: string;
        name: string;
        birth: string;
        purpose: string;
        sex: string;
        isCertified: boolean;
        warning: string;
        count?: number
}